Factor visibility calculation out of FogUtil
A new class `VisibilityProblem` is responsible for collecting blocking segments and generating a visibility polygon from
them. No meaningful changes have been made to the algorithm itself, though it does support some new conveniences such as
progressively building up the problem set before solving.
